refactor(homepage): for logic of join
